wip: fixing race-conditions when using async and sync simultaneously
